requirements.txt
discord.py
jishaku
humanize
